requirements.txt
numpy<2
streamlit
pandas
scikit-learn
scikit-surprise
